+++ src/main/java/net/spell_engine/compat/CompatFeatures.java
@@ -8,5 +8,6 @@ public class CompatFeatures {
         ContainerCompat.init();
         TrinketsCompat.init();
         CombatRollCompat.init();
+        MeleeCompat.init();
     }
 }